requirements.txt
python-telegram-bot
nc_py_api